requirements/base.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.1
whitenoise==6.8.2
